updated sinusoids~ to 64-bit

+++ src/sinusoids~/noise-table.h
@@ -1,3 +1,3 @@
-extern float NoiseTable[];
+extern double NoiseTable[];
 
 extern int NTS(void); // noise table size in floats
